fixed compilation errors on SFML, Qt and test

+++ libvpvl2/test/SceneTest.cc
@@ -29,7 +29,7 @@ using namespace vpvl2::extensions::icu4c;
 struct MockResolver : IApplicationContext::FunctionResolver {
     bool hasExtension(const char * /* name */) const { return false; }
     void *resolveSymbol(const char * /* name */) const { return 0; }
-    float query(QueryType /* type */) const { return 0; }
+    int query(QueryType /* type */) const { return 0; }
 } g_resolver;
 
 TEST(SceneTest, AddModel)
